mindflow/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
openai==1.3.0
streamlit==1.31.0
networkx==3.2.1
pydot==1.4.2
graphviz==0.20.1
requests==2.31.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
pydantic==2.5.3
